F Added readme, added SDP file, improved configuration, added comments

+++ src/RTSPServerFFmpeg/Program.cs
@@ -2,6 +2,7 @@
 using SharpRTSPServer;
 using System;
 using System.Diagnostics;
+using System.IO;
 using System.Text;
 using System.Threading;
 
@@ -11,37 +12,59 @@ ushort port = ushort.Parse(config["Port"]);
 string userName = config["UserName"];
 string password = config["Password"];
 
-string ffmpegPath = config["FFmpegPath"];
-string ffmpegArgs = config["FFmpegArgs"];
-string videoUri = config["VideoUri"];
-string audioUri = config["AudioUri"];
+string ffmpegPath = config["FFmpegPath"]; // path to ffmpeg.exe
+string ffmpegArgs = config["FFmpegArgs"]; // Arguments that will be passed to the ffmpeg process
+string videoUri = config["VideoUri"]; // RTP video URI
+string audioUri = config["AudioUri"]; // RTP audio URI
+string sdpFile = config["SDPFile"]; // SDP file path (Optional in case ffmpegPath and ffmpegArgs are not specified. You have to launch ffmpeg before starting the server.)
 
-// ffmpeg.exe -re -stream_loop -1 -i frag_bunny.mp4 -vcodec copy -an -f rtp rtp://127.0.0.1:11111 -vn -acodec copy -f rtp rtp://127.0.0.1:11113
-ProcessStartInfo info = new ProcessStartInfo();
-info.FileName = ffmpegPath;
-info.Arguments = ffmpegArgs;
-info.RedirectStandardOutput = true;
-info.UseShellExecute = false;
-
-string lastLine = null;
 SemaphoreSlim semaphore = new SemaphoreSlim(0);
-StringBuilder sdp = new StringBuilder();
-var process = Process.Start(info);
-process.OutputDataReceived += Process_OutputDataReceived;
-process.BeginOutputReadLine();
+StringBuilder sdpBuilder = new StringBuilder();
+ProcessStartInfo info = new ProcessStartInfo();
+Process process = null;
+string lastLine = null;
+string sdp = null;
 
-// wait until the SDP is read
-semaphore.Wait();
+if (!string.IsNullOrEmpty(ffmpegPath) && !string.IsNullOrEmpty(ffmpegArgs))
+{
+    // launch ffmpeg, parse the output and start streaming
+    // ffmpeg.exe -re -stream_loop -1 -i frag_bunny.mp4 -vcodec copy -an -f rtp rtp://127.0.0.1:11111 -vn -acodec copy -f rtp rtp://127.0.0.1:11113
+    info.FileName = ffmpegPath;
+    info.Arguments = ffmpegArgs;
+    info.RedirectStandardOutput = true;
+    info.UseShellExecute = false;
+
+    process = Process.Start(info);
+    process.OutputDataReceived += Process_OutputDataReceived;
+    process.BeginOutputReadLine();
+
+    // wait until the SDP is read
+    semaphore.Wait();
+
+    sdp = sdpBuilder.ToString();
+}
+else if(!string.IsNullOrEmpty(sdpFile))
+{
+    // optionally, read SDP from a file
+    sdp = File.ReadAllText(sdpFile);
+}
+else
+{
+    throw new Exception("Invalid configuration! Either ffmpegPath and ffmpegArgs, or SDPFile must be specified!");
+}
+
+if (string.IsNullOrEmpty(videoUri) && string.IsNullOrEmpty(audioUri))
+    throw new Exception("Invalid configuration! Either VideoUri, AudioUri or both must be specified!");
 
 using (var server = new RTSPServer(port, userName, password))
 {
     if (!string.IsNullOrEmpty(videoUri))
-        server.AddVideoTrack(new ProxyTrack(0, videoUri));
+        server.AddVideoTrack(new ProxyTrack(ProxyTrackType.Video, videoUri));
 
     if (!string.IsNullOrEmpty(audioUri))
-        server.AddAudioTrack(new ProxyTrack(1, audioUri));
+        server.AddAudioTrack(new ProxyTrack(ProxyTrackType.Audio, audioUri));
 
-    server.OverrideSDP(sdp.ToString(), true);
+    server.OverrideSDP(sdp, true);
 
     try
     {
@@ -57,10 +80,13 @@ using (var server = new RTSPServer(port, userName, password))
     Console.WriteLine("Press any key to exit");
     while (!Console.KeyAvailable)
     {
-        System.Threading.Thread.Sleep(250); 
+        Thread.Sleep(250); 
     }
 
-    process.Kill();
+    if (process != null)
+    {
+        process.Kill();
+    }
 }
 
 void Process_OutputDataReceived(object sender, DataReceivedEventArgs e)
@@ -75,6 +101,6 @@ void Process_OutputDataReceived(object sender, DataReceivedEventArgs e)
 
     if (!string.IsNullOrEmpty(e.Data) && !e.Data.StartsWith("SDP:"))
     {
-        sdp.AppendLine(e.Data);
+        sdpBuilder.AppendLine(e.Data);
     }
 }
